+++ CopilotKit/examples/next-openai/src/app/layout.tsx
@@ -2,13 +2,16 @@ import "../styles/globals.css";
 import "@copilotkit/react-ui/styles.css";
 import "@copilotkit/react-textarea/styles.css";
 import { ServiceAdapterSelector } from "./components/ServiceAdapterSelector";
+import { Suspense } from "react";
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en" className="bg-zinc-900">
       <body>
-        {children}
-        <ServiceAdapterSelector />
+        <Suspense fallback={<div>Loading...</div>}>
+          {children}
+          <ServiceAdapterSelector />
+        </Suspense>
       </body>
     </html>
   );