+++ CopilotKit/packages/react-textarea/src/hooks/make-autosuggestions-function/use-make-standard-autosuggestions-function.tsx
@@ -103,7 +103,7 @@ export function useMakeStandardAutosuggestionFunction(
           if (abortSignal.aborted) {
             break;
           }
-          if (message instanceof TextMessage) {
+          if (message.isTextMessage()) {
             result += message.content;
           }
         }