+++ CopilotKit/examples/next-pages-router/CHANGELOG.md
@@ -1,5 +1,17 @@
 # next-pages-router
 
+## 1.3.7
+
+### Patch Changes
+
+- Fix libraries for React 19 and Next.js 15 support
+- Updated dependencies
+  - @copilotkit/react-core@1.3.7
+  - @copilotkit/react-textarea@1.3.7
+  - @copilotkit/react-ui@1.3.7
+  - @copilotkit/runtime@1.3.7
+  - @copilotkit/shared@1.3.7
+
 ## 1.3.6
 
 ### Patch Changes