+++ CopilotKit/examples/node-http/CHANGELOG.md
@@ -1,5 +1,14 @@
 # node
 
+## 1.3.7
+
+### Patch Changes
+
+- Fix libraries for React 19 and Next.js 15 support
+- Updated dependencies
+  - @copilotkit/runtime@1.3.7
+  - @copilotkit/shared@1.3.7
+
 ## 1.3.6
 
 ### Patch Changes